@@ -1,17 +1,17 @@
 module github.com/bufbuild/httplb
 
-go 1.21
+go 1.23.0
 
 require (
 	github.com/jonboulle/clockwork v0.4.0
 	github.com/stretchr/testify v1.8.4
-	golang.org/x/net v0.33.0
-	golang.org/x/sync v0.10.0
+	golang.org/x/net v0.36.0
+	golang.org/x/sync v0.11.0
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	golang.org/x/text v0.21.0 // indirect
+	golang.org/x/text v0.22.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
