@@ -1,4 +1,4 @@
-// Copyright 2023 Buf Technologies, Inc.
+// Copyright 2023-2025 Buf Technologies, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -82,6 +82,7 @@ type leastLoadedRandom struct {
 	rng *rand.Rand
 }
 
+//nolint:recvcheck // mix of pointer and non-pointer receiver methods is intentional
 type leastLoadedConnHeap []*leastLoadedConnItem
 
 type leastLoadedConnItem struct {
