@@ -1,4 +1,4 @@
-// Copyright 2023 Buf Technologies, Inc.
+// Copyright 2023-2025 Buf Technologies, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -242,5 +242,5 @@ type dummyConn struct {
 }
 
 func connID(cn conn.Conn) string {
-	return cn.(dummyConn).id
+	return cn.(dummyConn).id //nolint:errcheck
 }
